Add blockchainRoutes and refactor code
