+++ server/src/models/Blockchain.mjs
@@ -1,14 +1,15 @@
-import { createHash } from "../utilities/crypto-lib.mjs";
-import Block from "./Block.mjs";
-import Transaction from "./Transaction.mjs";
+import { createHash } from '../utilities/cryptoLib.mjs';
+import Block from './Block.mjs';
+import Transaction from './Transaction.mjs';
 
 export default class Blockchain {
   constructor() {
     this.chain = [];
     this.memberNodes = [];
     this.pendingTransactions = [];
-
     this.nodeUrl = process.argv[3];
+
+    this.createBlock(Date.now(), '0', '0', [], 1500, process.env.DIFFICULTY);
   }
 
   createBlock(
@@ -48,67 +49,62 @@ export default class Blockchain {
   }
 
   hashBlock(timestamp, prevBlockHash, currBlockHash, nonce, difficulty) {
-    const stringToHash = timestamp.toString() +
+    const stringToHash =
+      timestamp.toString() +
       timestamp.toString() +
       prevBlockHash +
       JSON.stringify(currBlockHash) +
       nonce +
       difficulty;
-    const hash = createHash()
-  }
- }
-
- validateChain(blockchain) {
-  let isValid = true;
-
-  for (let i = 1; i < blockchain.length; i++) {
-    const currBlock = blockchain[i];
-    const prevBlock = blockchain[i - 1];
-
-    const hash = this.hashBlock(
-      currBlock.timestamp,
-      prevBlock.currentBlockHash,
-      currBlock.data
-    );
-
-    if (hash !== currBlock.currBlockHash) isValid = false;
-    if (currBlock.prevBlockHash !== prevBlock.currBlockHash)
-      isValid = false;
+    const hash = createHash(stringToHash);
   }
 
-  return isValid;
-}
+  validateChain(blockchain) {
+    let isValid = true;
 
-proofOfWork(prevBlockHash, data) {
-  const lastBlock = this.getLastBlock();
-  let difficulty, hash, timestamp, nonce = 0;
+    for (let i = 1; i < blockchain.length; i++) {
+      const currBlock = blockchain[i];
+      const prevBlock = blockchain[i - 1];
 
-  do {
-    nonce++;
-    timestamp = Date.now();
+      const hash = this.hashBlock(
+        currBlock.timestamp,
+        prevBlock.currentBlockHash,
+        currBlock.data
+      );
 
-    difficulty = this.difficultyAdjustment(lastBlock, timestamp);
-    hash = this.hashBlock(
+      if (hash !== currBlock.currBlockHash) isValid = false;
+      if (currBlock.prevBlockHash !== prevBlock.currBlockHash) isValid = false;
+    }
+
+    return isValid;
+  }
+
+  proofOfWork(prevBlockHash, data) {
+    const lastBlock = this.getLastBlock();
+    let difficulty,
+      hash,
       timestamp,
-      prevBlockHash,
-      data,
-      nonce,
-      difficulty
-    );
-  } while (hash.substring(0, difficulty) !== '0'.repeat(difficulty));
+      nonce = 0;
 
-  return { nonce, difficulty, timestamp };
+    do {
+      nonce++;
+      timestamp = Date.now();
+
+      difficulty = this.difficultyAdjustment(lastBlock, timestamp);
+      hash = this.hashBlock(timestamp, prevBlockHash, data, nonce, difficulty);
+    } while (hash.substring(0, difficulty) !== '0'.repeat(difficulty));
+
+    return { nonce, difficulty, timestamp };
+  }
+
+  difficultyAdjustment(lastBlock, timestamp) {
+    const MINE_RATE = process.env.MINE_RATE;
+    let { difficulty } = lastBlock;
+
+    if (difficulty < 1) return 1;
+
+    return timestamp - lastBlock.timestamp > MINE_RATE
+      ? +difficulty - 1
+      : +difficulty + 1;
+  }
 }
-
-difficultyAdjustment(lastBlock, timestamp) {
-  const MINE_RATE = process.env.MINE_RATE;
-  let { difficulty } = lastBlock;
-
-  if (difficulty < 1) return 1;
-
-  return timestamp - lastBlock.timestamp > MINE_RATE
-    ? +difficulty - 1
-    : +difficulty + 1;
-}
-
-
